# openapi-spec.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/v1/notifications": {
            "get": {
                "operationId": "NotificationsController_getNotifications",
                "summary": "Get paginated notifications",
                "parameters": [
                    {
                        "name": "page",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "minimum": 1,
                            "default": 1,
                            "type": "number"
                        }
                    },
                    {
                        "name": "limit",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "minimum": 1,
                            "maximum": 100,
                            "default": 20,
                            "type": "number"
                        }
                    },
                    {
                        "name": "isRead",
                        "required": false,
                        "in": "query",
                        "description": "Filter by read status",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "type",
                        "required": false,
                        "in": "query",
                        "description": "Filter by notification type",
                        "schema": {
                            "enum": [
                                "system",
                                "promotion",
                                "wallet",
                                "game",
                                "leaderboard",
                                "bet"
                            ],
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Notifications"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/notifications/unread-count": {
            "get": {
                "operationId": "NotificationsController_getUnreadCount",
                "summary": "Get count of unread notifications",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Notifications"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/notifications/{id}/read": {
            "patch": {
                "operationId": "NotificationsController_markAsRead",
                "summary": "Mark a single notification as read",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "Notification ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Notifications"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/notifications/mark-all-read": {
            "post": {
                "operationId": "NotificationsController_markAllAsRead",
                "summary": "Mark all notifications as read",
                "parameters": [],
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "Notifications"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/notifications/{id}": {
            "delete": {
                "operationId": "NotificationsController_deleteNotification",
                "summary": "Delete a notification",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "Notification ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Notifications"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/notifications/preferences": {
            "get": {
                "operationId": "NotificationsController_getPreferences",
                "summary": "Get notification preferences",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Notifications"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            },
            "patch": {
                "operationId": "NotificationsController_updatePreferences",
                "summary": "Update notification preferences",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateNotificationPreferencesDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Notifications"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
...
        "/api/v1/jackpots/recent-wins": {
            "get": {
                "operationId": "JackpotController_getRecentWins",
                "summary": "Get recent jackpot wins",
                "parameters": [
                    {
                        "name": "gameId",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "jackpotId",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "page",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Jackpots"
                ]
            }
        },
        "/api/v1/jackpots/type/{type}": {
            "get": {
                "operationId": "JackpotController_getJackpotByType",
                "summary": "Get jackpot by type (mini, minor, major, grand)",
                "parameters": [
                    {
                        "name": "type",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Jackpots"
                ]
            }
        },
        "/api/v1/jackpots/{id}": {
            "get": {
                "operationId": "JackpotController_getJackpotById",
                "summary": "Get jackpot details by ID",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Jackpots"
                ]
            }
        },
...
        "/api/v1/media/assets/{key}": {
            "get": {
                "operationId": "MediaController_getAsset",
                "summary": "Get asset by key",
                "parameters": [
                    {
                        "name": "key",
                        "required": true,
                        "in": "path",
                        "description": "Asset key (e.g., game:gates-of-olympus)",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns the asset details"
                    },
                    "404": {
                        "description": "Asset not found"
                    }
                },
                "tags": [
                    "Media"
                ]
            }
        },
        "/api/v1/media/url/{key}": {
            "get": {
                "operationId": "MediaController_getAssetUrl",
                "summary": "Get asset URL by key (returns placeholder if not found)",
                "parameters": [
                    {
                        "name": "key",
                        "required": true,
                        "in": "path",
                        "description": "Asset key (e.g., game:gates-of-olympus)",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns the asset URL or placeholder"
                    }
                },
                "tags": [
                    "Media"
                ]
            }
        },
        "/api/v1/media/urls": {
            "get": {
                "operationId": "MediaController_getAssetUrls",
                "summary": "Get multiple asset URLs by keys",
                "parameters": [
                    {
                        "name": "keys",
                        "required": true,
                        "in": "query",
                        "description": "Comma-separated asset keys",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns a map of keys to URLs"
                    }
                },
                "tags": [
                    "Media"
                ]
            }
        },
        "/api/v1/media/category/{slug}": {
            "get": {
                "operationId": "MediaController_getAssetsByCategory",
                "summary": "Get all assets in a category",
                "parameters": [
                    {
                        "name": "slug",
                        "required": true,
                        "in": "path",
                        "description": "Category slug (e.g., game-thumbnails)",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns all assets in the category"
                    }
                },
                "tags": [
                    "Media"
                ]
            }
        }
    },
    "info": {
        "title": "WBC 2026 API",
        "description": "WBC 2026 Social Casino Platform API Documentation",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "JWT-auth": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http",
                "name": "JWT",
                "description": "Enter JWT token",
                "in": "header"
            }
        },
        "schemas": {
...
            "UpdateNotificationPreferencesDto": {
                "type": "object",
                "properties": {
                    "emailPromotions": {
                        "type": "boolean",
                        "description": "Receive promotional emails"
                    },
                    "emailTransactions": {
                        "type": "boolean",
                        "description": "Receive transaction emails"
                    },
                    "pushEnabled": {
                        "type": "boolean",
                        "description": "Enable push notifications"
                    },
                    "pushPromotions": {
                        "type": "boolean",
                        "description": "Receive promotional push notifications"
                    },
                    "pushTransactions": {
                        "type": "boolean",
                        "description": "Receive transaction push notifications"
                    },
                    "smsEnabled": {
                        "type": "boolean",
                        "description": "Enable SMS notifications"
                    }
                }
            }
        }
    }
}